Module aws_sdk_budgets::error

Modules

access_denied_exception

See AccessDeniedException

creation_limit_exceeded_exception

See CreationLimitExceededException

duplicate_record_exception

See DuplicateRecordException

expired_next_token_exception

See ExpiredNextTokenException

internal_error_exception

See InternalErrorException

invalid_next_token_exception

See InvalidNextTokenException

invalid_parameter_exception

See InvalidParameterException

not_found_exception

See NotFoundException

resource_locked_exception

See ResourceLockedException

Structs

AccessDeniedException

You are not authorized to use this operation with the given parameters.

CreateBudgetActionError

CreateBudgetError

CreateNotificationError

CreateSubscriberError

CreationLimitExceededException

You've exceeded the notification or subscriber limit.

DeleteBudgetActionError

DeleteBudgetError

DeleteNotificationError

DeleteSubscriberError

DescribeBudgetActionError

DescribeBudgetActionHistoriesError

DescribeBudgetActionsForAccountError

DescribeBudgetActionsForBudgetError

DescribeBudgetError

DescribeBudgetPerformanceHistoryError

DescribeBudgetsError

DescribeNotificationsForBudgetError

DescribeSubscribersForNotificationError

DuplicateRecordException

The budget name already exists. Budget names must be unique within an account.

ExecuteBudgetActionError

ExpiredNextTokenException

The pagination token expired.

InternalErrorException

An error on the server occurred during the processing of your request. Try again later.

InvalidNextTokenException

The pagination token is invalid.

InvalidParameterException

An error on the client occurred. Typically, the cause is an invalid input value.

NotFoundException

We can’t locate the resource that you specified.

ResourceLockedException

The request was received and recognized by the server, but the server rejected that

UpdateBudgetActionError

UpdateBudgetError

UpdateNotificationError

UpdateSubscriberError

Enums

CreateBudgetActionErrorKind

CreateBudgetErrorKind

CreateNotificationErrorKind

CreateSubscriberErrorKind

DeleteBudgetActionErrorKind

DeleteBudgetErrorKind

DeleteNotificationErrorKind

DeleteSubscriberErrorKind

DescribeBudgetActionErrorKind

DescribeBudgetActionHistoriesErrorKind

DescribeBudgetActionsForAccountErrorKind

DescribeBudgetActionsForBudgetErrorKind

DescribeBudgetErrorKind

DescribeBudgetPerformanceHistoryErrorKind

DescribeBudgetsErrorKind

DescribeNotificationsForBudgetErrorKind

DescribeSubscribersForNotificationErrorKind

ExecuteBudgetActionErrorKind

UpdateBudgetActionErrorKind

UpdateBudgetErrorKind

UpdateNotificationErrorKind

UpdateSubscriberErrorKind