Module aws_sdk_budgets::model

Modules

action

See Action

action_history

See ActionHistory

action_history_details

See ActionHistoryDetails

action_threshold

See ActionThreshold

budget

See Budget

budget_performance_history

See BudgetPerformanceHistory

budgeted_and_actual_amounts

See BudgetedAndActualAmounts

calculated_spend

See CalculatedSpend

cost_types

See CostTypes

definition

See Definition

iam_action_definition

See IamActionDefinition

notification

See Notification

notification_with_subscribers

See NotificationWithSubscribers

scp_action_definition

See ScpActionDefinition

spend

See Spend

ssm_action_definition

See SsmActionDefinition

subscriber

See Subscriber

time_period

See TimePeriod

Structs

Action

ActionHistory

ActionHistoryDetails

ActionThreshold

Budget

Represents the output of the CreateBudget operation. The content consists of the detailed metadata and data file information, and the current status of the budget object.

BudgetPerformanceHistory

A history of the state of a budget at the end of the budget's specified time period.

BudgetedAndActualAmounts

The amount of cost or usage that you created the budget for, compared to your actual costs or usage.

CalculatedSpend

The spend objects that are associated with this budget. The actualSpend tracks how much you've used, cost, usage, RI units, or Savings Plans units and the forecastedSpend tracks how much you are predicted to spend based on your historical usage profile.

CostTypes

The types of cost that are included in a COST budget, such as tax and subscriptions.

Definition

IamActionDefinition

Notification

A notification that is associated with a budget. A budget can have up to ten notifications.

NotificationWithSubscribers

A notification with subscribers. A notification can have one SNS subscriber and up to 10 email subscribers, for a total of 11 subscribers.

ScpActionDefinition

Spend

The amount of cost or usage that is measured for a budget.

SsmActionDefinition

Subscriber

The subscriber to a budget notification. The subscriber consists of a subscription type and either an Amazon SNS topic or an email address.

TimePeriod

The period of time that is covered by a budget. The period has a start date and an end date. The start date must come before the end date. There are no restrictions on the end date.

Enums

ActionStatus

ActionSubType

ActionType

ApprovalModel

BudgetType

The type of a budget. It must be one of the following types:

ComparisonOperator

The comparison operator of a notification. Currently the service supports the following operators:

EventType

ExecutionType

NotificationState

NotificationType

The type of a notification. It must be ACTUAL or FORECASTED.

SubscriptionType

The subscription type of the subscriber. It can be SMS or EMAIL.

ThresholdType

The type of threshold for a notification.

TimeUnit

The time unit of the budget, such as MONTHLY or QUARTERLY.