Module aws_sdk_budgets::operation

Structs

CreateBudget

Creates a budget and, if included, notifications and subscribers.

CreateBudgetAction

CreateNotification

Creates a notification. You must create the budget before you create the associated notification.

CreateSubscriber

Creates a subscriber. You must create the associated budget and notification before you create the subscriber.

DeleteBudget

Deletes a budget. You can delete your budget at any time.

DeleteBudgetAction

DeleteNotification

Deletes a notification.

DeleteSubscriber

Deletes a subscriber.

DescribeBudget

Describes a budget.

DescribeBudgetAction

DescribeBudgetActionHistories

DescribeBudgetActionsForAccount

DescribeBudgetActionsForBudget

DescribeBudgetPerformanceHistory

Describes the history for DAILY, MONTHLY, and QUARTERLY budgets. Budget history isn't available for ANNUAL budgets.

DescribeBudgets

Lists the budgets that are associated with an account.

DescribeNotificationsForBudget

Lists the notifications that are associated with a budget.

DescribeSubscribersForNotification

Lists the subscribers that are associated with a notification.

ExecuteBudgetAction

UpdateBudget

Updates a budget. You can change every part of a budget except for the budgetName and the calculatedSpend. When you modify a budget, the calculatedSpend drops to zero until AWS has new usage data to use for forecasting.

UpdateBudgetAction

UpdateNotification

Updates a notification.

UpdateSubscriber

Updates a subscriber.